+++ packages/orq-rc/src/models/components/evaluatorresponsepython.ts
@@ -32,8 +32,8 @@ export const EvaluatorResponsePython$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-08-19T10:58:05.439Z"),
-  updated: z.string().default("2026-08-19T10:58:05.439Z"),
+  created: z.string().default("2026-08-19T12:36:38.391Z"),
+  updated: z.string().default("2026-08-19T12:36:38.391Z"),
   updated_by_id: z.nullable(z.string()).optional(),
   project_id: z.string().optional(),
   guardrail_config: z.any().optional(),