+++ packages/orq-rc/src/models/operations/createannotationqueue.ts
@@ -151,7 +151,7 @@ export const CreateAnnotationQueueResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-08-19T10:58:01.279Z",
+    "2026-08-19T12:36:36.185Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {