+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3894,7 +3894,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01M0CTRJQREWTV3JYTHEMNFSQS"),
+  id: z.string().default("01M0D0CE19968FZEMCHCCTR6X3"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),