+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -3953,7 +3953,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01M0CTRJRQYN7A7FST0KEPV5WM"),
+  id: z.string().default("01M0D0CE2ZST3GCBE2XE4V8VRK"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>