+++ packages/orq-rc/RELEASES.md
@@ -11739,3 +11739,13 @@ Based on:
 - [typescript v4.14.0-rc.37] packages/orq-rc
 ### Releases
 - [NPM v4.14.0-rc.37] https://www.npmjs.com/package/@orq-ai/node/v/4.14.0-rc.37 - packages/orq-rc
+
+## 2026-08-19 12:39:59
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.795.0 (2.931.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.14.0-rc.38] packages/orq-rc
+### Releases
+- [NPM v4.14.0-rc.38] https://www.npmjs.com/package/@orq-ai/node/v/4.14.0-rc.38 - packages/orq-rc